[Refactor] Split table creation logics to different places
Use abstract factory design pattern split the table creation
to different places based on the table engine;
