+++ fe/fe-core/src/main/java/com/starrocks/server/OlapTableFactory.java
@@ -1,0 +1,36 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package com.starrocks.server;
+
+import com.starrocks.catalog.Database;
+import com.starrocks.catalog.Table;
+import com.starrocks.common.DdlException;
+import com.starrocks.sql.ast.CreateTableStmt;
+
+import javax.validation.constraints.NotNull;
+
+public class OlapTableFactory implements AbstractTableFactory {
+    public static final OlapTableFactory INSTANCE = new OlapTableFactory();
+
+    private OlapTableFactory() {
+    }
+
+    @Override
+    @NotNull
+    public Table createTable(LocalMetastore metastore, Database database, CreateTableStmt stmt) throws DdlException {
+        // TODO: move createOlapOrLakeTable() to this class.
+        return metastore.createOlapOrLakeTable(database, stmt);
+    }
+}
