+++ fe/fe-core/src/test/java/com/starrocks/catalog/HiveTableTest.java
@@ -42,8 +42,9 @@ import com.starrocks.connector.hive.HiveMetastoreTest;
 import com.starrocks.qe.ConnectContext;
 import com.starrocks.server.GlobalStateMgr;
 import com.starrocks.server.MetadataMgr;
-import com.starrocks.server.TableFactory;
+import com.starrocks.server.TableFactoryProvider;
 import com.starrocks.sql.ast.CreateTableStmt;
+import com.starrocks.sql.common.EngineType;
 import com.starrocks.utframe.StarRocksAssert;
 import com.starrocks.utframe.UtFrameUtils;
 import mockit.Expectations;
@@ -74,6 +75,10 @@ public class HiveTableTest {
                 "\"type\"  =  \"hive\", \"hive.metastore.uris\"  =  \"thrift://127.0.0.1:9083\")");
         starRocksAssert.withDatabase("db");
         hiveClient = new HiveMetastoreTest.MockedHiveMetaClient();
+    }
+
+    com.starrocks.catalog.Table createTable(CreateTableStmt stmt) throws DdlException {
+        return TableFactoryProvider.getFactory(EngineType.HIVE.name()).createTable(null, null, stmt);
     }
 
     @Test
@@ -108,7 +113,7 @@ public class HiveTableTest {
         String createTableSql = "create external table db.hive_tbl (col1 int, col2 int) engine=hive properties " +
                 "(\"resource\"=\"hive0\", \"database\"=\"db0\", \"table\"=\"table0\")";
         CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(createTableSql, connectContext);
-        com.starrocks.catalog.Table table = TableFactory.createTable(createTableStmt, com.starrocks.catalog.Table.TableType.HIVE);
+        com.starrocks.catalog.Table table = createTable(createTableStmt);
 
         Assert.assertTrue(table instanceof HiveTable);
         HiveTable hiveTable = (HiveTable) table;
@@ -128,7 +133,7 @@ public class HiveTableTest {
         String createTableSql = "create external table nodb.hive_tbl (col1 int, col2 int) engine=hive properties " +
                 "(\"resource\"=\"hive0\", \"table\"=\"table0\")";
         CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(createTableSql, connectContext);
-        com.starrocks.catalog.Table table = TableFactory.createTable(createTableStmt, com.starrocks.catalog.Table.TableType.HIVE);
+        com.starrocks.catalog.Table table = createTable(createTableStmt);
         Assert.fail("No exception throws.");
     }
 
@@ -137,7 +142,7 @@ public class HiveTableTest {
         String createTableSql = "create external table nodb.hive_tbl (col1 int, col2 int) engine=hive properties " +
                 "(\"resource\"=\"hive0\")";
         CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(createTableSql, connectContext);
-        com.starrocks.catalog.Table table = TableFactory.createTable(createTableStmt, com.starrocks.catalog.Table.TableType.HIVE);
+        com.starrocks.catalog.Table table = createTable(createTableStmt);
         Assert.fail("No exception throws.");
     }
 
@@ -146,7 +151,7 @@ public class HiveTableTest {
         String createTableSql = "create external table db.hive_tbl (col1 int, col2 int) engine=hive properties " +
                 "(\"resource\"=\"not_exist_reousrce\", \"database\"=\"db0\", \"table\"=\"table0\")";
         CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(createTableSql, connectContext);
-        com.starrocks.catalog.Table table = TableFactory.createTable(createTableStmt, com.starrocks.catalog.Table.TableType.HIVE);
+        com.starrocks.catalog.Table table = createTable(createTableStmt);
         Assert.fail("No exception throws.");
     }
 
@@ -155,7 +160,7 @@ public class HiveTableTest {
         String createTableSql = "create external table db.hive_tbl (col1 int, col2 int) engine=hive properties " +
                 "(\"database\"=\"db0\", \"table\"=\"table0\")";
         CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(createTableSql, connectContext);
-        com.starrocks.catalog.Table table = TableFactory.createTable(createTableStmt, com.starrocks.catalog.Table.TableType.HIVE);
+        com.starrocks.catalog.Table table = createTable(createTableStmt);
         Assert.fail("No exception throws.");
     }
 
@@ -177,7 +182,7 @@ public class HiveTableTest {
         String createTableSql = "create external table db.hive_tbl (col1 int, not_exist int) engine=hive properties " +
                 "(\"resource\"=\"hive0\", \"database\"=\"db0\", \"table\"=\"table0\")";
         CreateTableStmt createTableStmt = (CreateTableStmt) UtFrameUtils.parseStmtWithNewParser(createTableSql, connectContext);
-        com.starrocks.catalog.Table table = TableFactory.createTable(createTableStmt, com.starrocks.catalog.Table.TableType.HIVE);
+        com.starrocks.catalog.Table table = createTable(createTableStmt);
         Assert.fail("No exception throws.");
     }
 }
